+++ src/app/shared/components/index.ts
@@ -4,3 +4,4 @@ export * from './quick-action-buttons/quick-action-buttons.component';
 export * from './installation-options/installation-options.component';
 export * from './setting-control/setting-control.component';
 export * from '../remote-config/cron-input/cron-input.component';
+export * from './json-editor/json-editor.component';